@@ -1,20 +1,22 @@
+
+import {map} from 'rxjs/operators';
+import {HttpClient} from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import {Http} from "@angular/http";
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+
+
 
 @Injectable()
 export class ConfigService {
   public radioSelected : string = "Autoclaves";
   config: any;
-  constructor(private http: Http) {
+  constructor(private http: HttpClient) {
     console.log('ctor for ConfigService called.')
   }
 
   load() {
     console.log('Inside Load');
     return new Promise((resolve) => {
-      this.http.get('./app.config.json').map(res => res.json())
+      this.http.get('./app.config.json')
         .subscribe(config => {
           console.log('Configuration loaded...........');
           this.config = config;
