@@ -1,6 +1,8 @@
+
+import {filter} from 'rxjs/operators';
 import { Component }                                from '@angular/core';
 import { Router, ActivatedRoute, NavigationEnd }    from '@angular/router';
-import 'rxjs/add/operator/filter';
+
 
 @Component({
     selector: 'breadcrumbs',
@@ -15,7 +17,7 @@ export class BreadcrumbsComponent {
     breadcrumbs: Array<Object>;
     constructor(private router:Router, private route:ActivatedRoute) {}
     ngOnInit(): void {
-        this.router.events.filter(event => event instanceof NavigationEnd).subscribe(event => {
+        this.router.events.pipe(filter(event => event instanceof NavigationEnd)).subscribe(event => {
             this.breadcrumbs = [];
             let currentRoute = this.route.root,
             url = '';
